Extract ReaderPostCardActionType and ReaderInterfaces to `:libs:reader`
Move ReaderPostCardActionType enum from ReaderCardUiState.kt into its
own file in libs:reader, unblocking ReaderInterfaces.java which also
moves.
