+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/ReaderCardUiState.kt
@@ -179,21 +179,3 @@ sealed class ReaderPostCardAction {
         override val onClicked: ((Long, Long, ReaderPostCardActionType) -> Unit)? = null
     ) : ReaderPostCardAction()
 }
-
-enum class ReaderPostCardActionType {
-    FOLLOW,
-    SITE_NOTIFICATIONS,
-    SHARE,
-    VISIT_SITE,
-    BLOCK_SITE,
-    BLOCK_USER,
-    LIKE,
-    BOOKMARK,
-    REBLOG,
-    COMMENTS,
-    REPORT_POST,
-    REPORT_USER,
-    TOGGLE_SEEN_STATUS,
-    SPACER_NO_ACTION,
-    READING_PREFERENCES,
-}
