+++ libs/reader/src/main/java/org/wordpress/android/ui/reader/discover/ReaderPostCardActionType.kt
@@ -1,0 +1,19 @@
+package org.wordpress.android.ui.reader.discover
+
+enum class ReaderPostCardActionType {
+    FOLLOW,
+    SITE_NOTIFICATIONS,
+    SHARE,
+    VISIT_SITE,
+    BLOCK_SITE,
+    BLOCK_USER,
+    LIKE,
+    BOOKMARK,
+    REBLOG,
+    COMMENTS,
+    REPORT_POST,
+    REPORT_USER,
+    TOGGLE_SEEN_STATUS,
+    SPACER_NO_ACTION,
+    READING_PREFERENCES,
+}
